% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_outcome_cleaning_function.r
\name{getAccuracy}
\alias{getAccuracy}
\title{get Accuracy
bias_res <- getAccuracy(sim_result)}
\usage{
getAccuracy(results)
}
\description{
get Accuracy
bias_res <- getAccuracy(sim_result)
}
