% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_generate_data.r
\name{ggenData}
\alias{ggenData}
\title{generally generate data}
\usage{
ggenData(
  n = 1000,
  Phi = 1,
  lam_x = 1,
  error_x = 0,
  lam_y = matrix(c(1, 0, 1, 1, 1, 2, 1, 3), nrow = 4, byrow = T),
  I1 = matrix(c(1, 0, 0, 1), nrow = 2),
  Beta = matrix(c(0, 0, 0, 0), nrow = 2, byrow = T),
  Gamma = matrix(c(0, 0), nrow = 2),
  Psi = matrix(c(1, 0.2, 0.2, 1), nrow = 2),
  error_y = diag(4),
  x = 0,
  eta_mu = cbind(c(5, -0.5))
)
}
\description{
generally generate data
}
