% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plot_table_function.r
\name{makePlot}
\alias{makePlot}
\title{make plots for results}
\usage{
makePlot(results, what = "Bias", font_size = 14, point_size = 2, alpha = 1)
}
\description{
make plots for results
}
\examples{
makePlot(bias_res, what = "RMSE", font_size = 12)
makePlot(power_res, what = "power", font_size = 12)
makePlot(power_res, what = "typeIerror", font_size = 12)
}
