% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plot_table_function.r
\name{makeTable}
\alias{makeTable}
\title{generate table
makeTable(bias_res, what = "RMSE", font_size = 12)
makeTable(power_res, what = "power", font_size = 12)
makeTable(power_res, what = "typeIerror", font_size = 12)}
\usage{
makeTable()
}
\description{
generate table
makeTable(bias_res, what = "RMSE", font_size = 12)
makeTable(power_res, what = "power", font_size = 12)
makeTable(power_res, what = "typeIerror", font_size = 12)
}
