% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_run_simulation_function.r
\name{runSimulation}
\alias{runSimulation}
\title{run simulation}
\usage{
runSimulation(condition_full, seed = 1)
}
\description{
run simulation
}
\examples{
sim_result <- runSimulation(condition_list)
}
