% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_trySimulation.r
\docType{package}
\name{trySimulation}
\alias{trySimulation}
\title{trySimulation: A package for practicing a simulation study with R.}
\description{
The trySimulation pacakge provides something.
}
\section{the functions}{

the functions something something
}

